Add performOperations
